Refactor thruster control
