+++ catkin_ws/src/CMakeLists.txt
@@ -1,0 +1,1 @@
+/opt/ros/kinetic/share/catkin/cmake/toplevel.cmake
