+++ catkin_ws/src/aquadrone_sim_demos/launch/empty.launch
@@ -1,6 +1,5 @@
 <launch>
-    <arg name="model_name" default="v2"/>
-    <arg name="control_file" default="$(arg model_name)_control.launch"/>
+    <arg name="model_name" default="v28"/>
 
     <include file="$(find aquadrone_sim_worlds)/launch/empty_world.launch">
         <arg name="paused" value="false"/>
@@ -14,6 +13,8 @@
         <arg name="model" value="aquadrone"/>
     </include>
 
-    <include file="$(find thruster_control)/launch/$(arg control_file)"/>
+    <include file="$(find thruster_control)/launch/sim_control.launch">
+        <arg name="model" value="$(arg model_name)"/>
+    </include>
 
 </launch>
