+++ catkin_ws/src/thruster_control/launch/sim_control.launch
@@ -1,4 +1,7 @@
 <launch>
+    <arg name="model"/>
+
+    <param name="model" value="$(arg model)"/>
 
     <node name="thruster_controller" pkg="thruster_control" type="sim_thruster_controller.py" output="screen"/>
     <node name="depth_pid" pkg="thruster_control" type="demo_depth_control.py" output="screen"/>
