+++ catkin_ws/src/thruster_control/nodes/sim_thruster_controller.py
@@ -2,16 +2,43 @@
 
 import rospy
 
-from thruster_control.thruster_collection_manager import ThrusterCollectionManager
-from thruster_control.thruster_interfaces import SimulatedThrusterInterface
-
+from thruster_control.thruster_configurations import V1Configuration, V2Configuration, V28Configuration
+from thruster_control.movement_command_collector import MovementCommandCollector
+from thruster_control.thruster_interfaces.sim_interface import SimulatedThrusterInterface
+from thruster_control.thruster_types import UUVSimThruster
+from thruster_control.thruster_control_manager import SimThrusterController
 
 
 if __name__ == "__main__":
     rospy.init_node('sim_thruster_controller')
 
-    thrusters = [SimulatedThrusterInterface('aquadrone', i) for i in range(0, 8)]
+    mcc = MovementCommandCollector()
 
-    tm = ThrusterCollectionManager(thrusters)
-    while not rospy.is_shutdown():
-        rospy.sleep(1)
+    # Assume V28 by default
+    model = rospy.get_param("model", "v28")
+    num = 8
+
+    if model == "v1":
+        config = V1Configuration()
+        num = 6
+    elif model == "v2":
+        config = V1Configuration()
+        num = 6
+    elif model == "v28":
+        config = V28Configuration()
+        num = 8
+    else:
+        print("Error: unknown model for controls: %s" % str(model))
+
+    config.initialize()
+
+    interface = SimulatedThrusterInterface("aquadrone", num)
+    interface.initialize()
+
+    th_spec = UUVSimThruster()
+    th_spec.initialize()
+    specs = [th_spec for i in range(0, 8)]
+
+    stc = SimThrusterController(config, mcc, interface, specs)
+    stc.run()
+
