+++ catkin_ws/src/thruster_control/src/thruster_control/geometry_helper.py
